fix:change some relation OneToMany to ManyToMany and feat:add the new entity GradeHistory and TrackSemesterCourse

+++ src/main/java/hei/school/exam/entity/Exam.java
@@ -16,7 +16,7 @@ public class Exam {
   @GeneratedValue(strategy = GenerationType.UUID)
   private UUID id;
 
-  private int coefficient;
+  private double coefficient;
   private LocalDate date;
 
   @ManyToOne private Course course;
